+++ src/wetterdienst/provider/knmi/observation/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (C) 2018-2025, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
+from wetterdienst.provider.knmi.observation.api import KnmiObservationRequest
+from wetterdienst.provider.knmi.observation.metadata import KnmiObservationMetadata
+
+__all__ = ["KnmiObservationMetadata", "KnmiObservationRequest"]